animation works
The scheduling part is a bit off, every access is plotted as its own
frame, but they should be parallel.
Colors are missing due to colorama docs beeing unavailable (no net in train).

We should get rid of the last remaining regex (see todos) and add tests/more
examples!

Thanks to Chris Liebing for the CLR Podcast,
it made coding in the hot+loud train possible :)